[FIX] Flux incremental emit issue has been fixed

+++ src/main/kotlin/com/example/stockservice/Review.kt
@@ -1,6 +1,7 @@
 package com.example.stockservice
 
+import java.math.BigDecimal
 import java.time.LocalDate
 
-data class Review( val reviewId:String,  val rating:Int,val customerProfileId:String, val title:String , val author:String,val reviewText:String ,
+data class Review( val reviewId:String,  val rating:BigDecimal,val customerProfileId:String, val title:String , val author:String,val reviewText:String ,
                    val reviewDate:LocalDate)
